+++ app/__init__.py
@@ -2,6 +2,11 @@ import logging, sys
 from . import cli, gui
 from .version import VERSION_STR
 
+class RunMode:
+    NONE = 0
+    GUI = 1
+    CLI = 2
+
 def run_app():
     '''
     Main entry point. Run this function to start the application.
@@ -11,9 +16,16 @@ def run_app():
     logging.basicConfig(level=logging.DEBUG, format='[%(levelname)s] %(message)s')
     logging.info(f"Running SteamJar v{VERSION_STR}")
 
-    # Run application
-    run_cli = sys.stdin and sys.stdin.isatty() and '--gui' not in sys.argv
-    if not run_cli:
-        run_cli = not gui.run()
-    if run_cli:
+    # Select run mode
+    mode = RunMode.GUI
+    if '--cli' in sys.argv: mode = RunMode.CLI
+    if '--gui' in sys.argv: mode = RunMode.GUI
+    
+    # Run in GUI mode
+    if mode == RunMode.GUI:
+        result = gui.run()
+        mode = RunMode.NONE if result else RunMode.CLI
+
+    # Run in CLI mode
+    if mode == RunMode.CLI:
         cli.run()